requirements.txt
alembic==1.14.1
babel==2.16.0
blinker==1.9.0
boto3==1.36.8
botocore==1.36.8
cachelib==0.13.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==5.5.2
Flask-Security-Too==5.5.2
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
msgspec==0.19.0
numpy==2.2.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
reportlab==4.2.5
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
speaklater==1.3
SQLAlchemy==2.0.37
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
XlsxWriter==3.2.2